% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agedata.R
\docType{data}
\name{agedata}
\alias{agedata}
\title{Continuous repeated measurement data}
\format{
A \code{tibble} with 6 columns which are :
\describe{
\item{age}{age of subject}
\item{agemiss}{age with missing value}
\item{1}{Observation on timepoint 1}
\item{2}{Observation on timepoint 2}
\item{3}{Observation on timepoint 3}
\item{4}{Observation on timepoint 4}}
}
\usage{
data(agedata)
}
\description{
dataset of observations made at various timepoints,variable age with 
no missing value and  agemiss with missing value
}
\keyword{datasets}
