% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countdata.R
\docType{data}
\name{countdata}
\alias{countdata}
\title{Repeated measurement data}
\format{
A \code{tibble} with 7 columns which are :
\describe{
\item{X1}{Observation on timepoint 1}
\item{X2}{Observation on timepoint 2}
\item{X3}{Observation on timepoint 3}
\item{X4}{Observation on timepoint 4}
\item{X1}{Observation on timepoint 2}
\item{X2}{Observation on timepoint 7}
\item{X4}{Observation on timepoint 14}}
}
\usage{
data(countdata)
}
\description{
dataset of observations made at various timepoints, variables take integer value.
first row refers timepoint
}
\keyword{datasets}
