% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_ipw}
\alias{print_ipw}
\title{print method for ipw}
\usage{
print_ipw(x, ...)
}
\arguments{
\item{x}{ipw object}

\item{...}{further argument can be passed}
}
\value{
print result for ipw object
}
\description{
print method for ipw
}
