% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_meanscore}
\alias{print_meanscore}
\title{print method for meanscore}
\usage{
print_meanscore(x, ...)
}
\arguments{
\item{x}{meanscore object}

\item{...}{further argument can be passed}
}
\value{
print result for meanscore object
}
\description{
print method for meanscore
}
