% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propdata.R
\docType{data}
\name{propdata}
\alias{propdata}
\title{Categorical,Continuous repeated measurement data}
\format{
A \code{tibble} with 17 columns which are :
\describe{
\item{X1...X6}{categorical data with values 0,1}
\item{X1..X11}{continuous data}}
}
\usage{
data(propdata)
}
\description{
dataset of observations made at various timepoints,first row refers probability
}
\keyword{datasets}
