% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatdata.R
\docType{data}
\name{repeatdata}
\alias{repeatdata}
\title{Continuous repeated measurement data}
\format{
A \code{tibble} with 3 columns which are :
\describe{
\item{2}{Observation on timepoint 2}
\item{4}{Observation on timepoint 4}
\item{14}{Observation on timepoint 14}}
}
\usage{
data(repeatdata)
}
\description{
dataset whose first row is the respective time points at which observations are taken.First row refers to time points.
}
\keyword{datasets}
