% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdata1.R
\docType{data}
\name{srdata1}
\alias{srdata1}
\title{protein data}
\format{
A dataframe with 164 rows and 9 columns
\describe{
\item{ID}{ID of subjects}
\item{Visit}{Number of times observations recorded}
\item{C6kine,.....,GFRalpha4}{These are covariates}}
}
\usage{
data(srdata1)
}
\description{
Repeated measurement dataset, for each id we have four visit observations
}
\examples{
data(srdata1)
}
\keyword{datasets}
