% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_ipw}
\alias{summary_ipw}
\title{summary method for ipw}
\usage{
summary_ipw(object, ...)
}
\arguments{
\item{object}{ipw object}

\item{...}{further argument can be passed}
}
\value{
summary of ipw object
}
\description{
summary method for ipw
}
