% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_meanscore}
\alias{summary_meanscore}
\title{summary method for meanscore}
\usage{
summary_meanscore(object, ...)
}
\arguments{
\item{object}{meanscore object}

\item{...}{further argument can be passed}
}
\value{
summary of meanscore object
}
\description{
summary method for meanscore
}
