% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateALpha.r
\name{updateALpha}
\alias{updateALpha}
\title{internal function for updating alpha}
\usage{
updateALpha(y, x, vfun, mu, w, phi, corstr, ni, mv = NULL, id, visit)
}
\arguments{
\item{y}{response value for GEE model}

\item{x}{model matrix for the GEE model}

\item{vfun}{variance function for the GLM}

\item{mu}{mu vector for the GLM}

\item{w}{weight matrix}

\item{phi}{scale parameter}

\item{corstr}{correlation structure}

\item{ni}{list of visits per subject}

\item{mv}{NULL}

\item{id}{id column}

\item{visit}{visit column}
}
\description{
internal function for updating alpha
}
\details{
arguments are from Fisher Scoring Algorithm
}
