% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateBeta.r
\name{updateBeta}
\alias{updateBeta}
\title{internal function for updating beta through Fisher Scoring}
\usage{
updateBeta(y, x, vfun, mu, w, D, Ralpha, beta)
}
\arguments{
\item{y}{response value for GEE model}

\item{x}{model matrix for the GEE model}

\item{vfun}{variance function for the GLM}

\item{mu}{mu vector for the GLM}

\item{w}{weight matrix}

\item{D}{derivation of the inverse link function}

\item{Ralpha}{correlation matrix}

\item{beta}{vector of beta value for GEE model}
}
\description{
internal function for updating beta through Fisher Scoring
}
