% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdatePhi.r
\name{UpdatePhi}
\alias{UpdatePhi}
\title{internal function for updating scale parameter}
\usage{
UpdatePhi(y, x, vfun, mu, w)
}
\arguments{
\item{y}{response value for GEE model}

\item{x}{model matrix for the GEE model}

\item{vfun}{variance function for the GLM}

\item{mu}{mu vector for the GLM}

\item{w}{weight matrix}
}
\description{
internal function for updating scale parameter
}
