% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSandW.r
\name{updateSandW}
\alias{updateSandW}
\title{internal function for sandwich estimator}
\usage{
updateSandW(y, x, vfun, mu, w, D, Ralpha, beta, hessmat, blockdiag)
}
\arguments{
\item{y}{response value for GEE model}

\item{x}{model matrix for the GEE model}

\item{vfun}{variance function for the GLM}

\item{mu}{mu vector for the GLM}

\item{w}{weight matrix}

\item{D}{derivation of the inverse link function}

\item{Ralpha}{correlation matrix}

\item{beta}{vector of beta value for GEE model}

\item{hessmat}{hessian matrix}

\item{blockdiag}{vector containing the dim of block matrix for block diagonal matrix}
}
\description{
internal function for sandwich estimator
}
\details{
arguments are required for obtaining Sandwich Estimator for variance matrix of regression coefficient of GEE model
}
